requirements.txt
python-telegram-bot==21.3
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
APScheduler==3.10.4
python-dotenv==1.0.1
